back/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-MySQLdb==2.0.0
numpy==1.26.4
pandas==2.2.2
tensorflow==2.17.0
pillow==10.4.0
python-dotenv==1.0.1
scikit-learn==1.5.1
werkzeug==3.0.4
